Adds support for DTASelect
DTASelect filtering task is included at the end of a Celery task chain. As with the Blazmass jobs, this is a PBS job wrapped in a Celery task with Celery's automatic retries.

Tasks/jobs are submitted using submit_and_check_job with a new dta_task flag set to True. (this disables the usual SQT file quality checks, since they don't apply here)